Added the SWT Client to the Tabris Server
The SWTClient has been added with the ID com.eclipsesource.tabris.swt
and a CSS file called swt.css. The CSS File is not complete, and
it will be extended as we move forward.

Currently the SWT client only supports the BrowserNavigation service
on the server. This is so the ControlsDemo will work. We may change
this in the future.

The TabrisSWTClientProvider was added to provide the SWTClient
to the system.

Change-Id: Ice0620f0c2e6cc73a6bbbeb72d16ad3c726cc615