Add a service to track the table item height
Because the SWT Client cannot set the item height to the exact size the
server wants, a callback was used to notify the server of the item
height for client tables. This callback was implemented as a
ClientService using a Call Operation. The call operation is
"setItemHeight" and it takes two parameters:
 - "itemHeight": the height of the items
 - "target": the widgetID of the table / tree widget
 
Change-Id: Ic0521958a62faee1f923b8e254371f983ed3a7b1