Server Side solution for #310: Add Pull-to-Refresh Widget
The widget is named RefreshComposite and is basically a Composite. It
can do everything a Composite can do plus 2 additional features.

1) You can set a message that should be displayed when you pull down the
widget. The protocol property is called "message"

2) You can add RefreshListeners which will be notified when the user
pulls down the widget and let it snap back. The protocol event type is
called "Refresh" and has no properties. A respecting listen operation
will also be sent.

Change-Id: I4e6f37cbc82e649acdadde9afb2f8be01e292ff0