Server Side Solution for #312: Add Support for native Dialogs
To create a native dialog a new type is introduced called ClientDialog.
A dialog can have a title, message, severity and 3 buttons max. The
buttons can only be used together with a button type which is OK, CANCEL
or CUSTOM. The client needs to decide where to place those buttons. A
button can be set together with a SWT listener to get notifications when
it was pressed.

To open and close a dialog the open and close methods on ClientDialog
has to be called. Because ClientDialog are non-blocking by default no
DialogUtil is needed here.

Change-Id: I090b552b6920cb8b841ca7364d93d4d02b756fe5