Enhances Tabris UI to enable UI interaction tracking
Page navigation could be tracked using TransistionListeners. This commit
adds a new listener type called ActionListener. It can be used to get
notification about action executions. As with TransistionListeners these
can be added to a UIConfiguration.

To gain Page and Action information at runtime the UIConfiguration was
modified. It's now possible to get a PageConfiguration or
ActionConfiguration for a Page or Action instance.

Change-Id: I3274d8123cd890bd16790fdf4e9191d31e7b1d9a