Set the padding for Table Headers in TabrisSWT
TableColumn padding is used in the calculation of Header Height. Since
we are only focused on Windows right now, we set the value to 9px, which
through trial and error (and a screen ruler) seems good.

This change also removes the font on table-cell since we should simply
inherit the default (16px Helvetica was wrong). Finally, the cell
padding was removed as this is not needed.