Fixes #327: Adds support for tracking services like Google Analytics
Adds a new bundle called com.eclipsesource.tabris.tracking. This bundle
provides an API for registering a "Tracker" into a "TabrisTracking"
instance. The tracking can be started by calling "start" with a
UIConfiguration.

One tracker was also implemented. The "GoogleAnalyticsTracker" which
hooks Google Analytics.

The actual tracking happens asynchronously with a scheduled executor
that dispatches "TrackingEvents" to the responsible "Trackers" every 5
seconds.

Change-Id: Ib33724acf5d7f1a3264dad22a169d6ec5b9f2d41