Improves Test Stability
Makes the tabris tests independent of the RAP test fixture to be more
stable for future changes. It adds junit rule called TabrisEnvironment
which encapsulates the whole tabris setup. It also allows the
dispatching of operations on remote objects. Besides this a MessageUtil
was added to make the verification if a message was send to the cleint
or not easier.

Change-Id: Ife0d372971a8d00160fc07ad40587fb2401b7914