Preparation for custom tracking events
To support cusotm events like ecommerce it's importend to be able to
share the Tacking instance in multiple sessions. For this reason the
dispatch queue need to be the same for every Tracking because it's not
guaranteed when a Tracking is created.

This change introduces the DispatchTask which just encapsulates the
Tracker instances and the TrackingEvent. This allows us to remove the
trackers field form the EventDispatcher and hold the dispatcher as a
real singleton (which is what we want).

Change-Id: Ibf5dd850c37fa1d83a2be0de37cc7ef6ec519443