Adds API for custom tracking
Besides attaching a UIConfiguration to a Tracking it's now allowed to
submit custom events like ecommerce and event. 

An event is jsut a simple action with an ID. This can be used to track
widget selection and so on.

An ecommerce event is represented by an Order. An Order can hold several
OrderItems. 

This change modifies the GoogleAnalyticsTracker and the ConsoleTracker
to make use of thos enew events. The PiwikTracker will follow later.


Change-Id: I20a81d90425be231ca60ee41d4d969e795ce902c