Adds a ClientService for CloudPush notifications
A new ClientService was introduced called CloudPush. The service accepts
a CloudPushListener. This instructs a client to register itself to
receive push notifications.

Afterwards the client has to send notifications to the server when the
registration for platform specific cloud push service was done. The
notifaction has to be of the type "Registered" and needs to contain one
property called "token" which should contain the registration id.

When the cloud push service triggers push notifications the client has
to send a notify operation of the type "MessageReceived" and needs to
contain one property called "message". The property "message" needs to
contain the whole message unparsed.

Errors are also handled by a notify operation of the type
"ErrorReceived".

Change-Id: Ic385398fc9d2217ce5195b38c9d62eca95c7046c