Use RAP file upload to handle camera image
In some cases (under Windows for example), the image taken by the camera
is very big. It becomes even bigger when we create a base64 from it and
use RAP UI request to send it to the server. By using standard file
upload process to the image file size does not matter. 

Change-Id: Ie462db13b4aee154fabdd9b5ee6405f3185e8218